Move findTile() from src/helpers/map.js to src/helpers/finders.js
