+++ src/helpers/map.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { PPTL, CPTL, PPCL } from '../constants';
+import { CPTL, PPCL } from '../constants';
 import { assert, assertGridInBounds } from '../utils/asserts';
 import { getTileProperty, tileHasProperty, tileIsType } from '../tiles';
 import { updateNTSprites, generatePermanentNTSprites } from '../draw/drawings';
@@ -170,35 +170,4 @@ export function getMapTraversabilityInCells(map) {
     updateNTSprites(xy.x, xy.y, mapTraversabilityCells);
   }
   return mapTraversabilityCells;
-}
-
-
-/*
- * Returns the position x and y (in pixels) of the first of the specified tile
- * types to appear starting in the top left corner and moving in a page-reading
- * fashion.
- *
- * Runtime: O(N^2)
- *
- * @param {(number | number[])} tiles - either a number representing a tileType,
- * or an array of such numbers
- */
-export function findTile(map, tileIds) {
-  assert(map, 'map is undefined');
-  assert(tileIds, 'tileIds is undefined');
-
-  // Force an array if the input is just one tile
-  const tileIdArray = [].concat(tileIds);
-
-  for (let x = 0, xl = map.length, yl = map[0].length; x < xl; x++) {
-    for (let y = 0; y < yl; y++) {
-      for (let i = 0; i < tileIdArray.length; i++) {
-        const tileId = tileIdArray[i];
-        if (map[x][y] === tileId) {
-          return { x: x * PPTL, y: y * PPTL };
-        }
-      }
-    }
-  }
-  throw new Error(`Unable to find tile: ${tileIds}`);
 }
